requirements.txt
numpy
tensorflow-gpu
easydict
pynrrd
pyyaml
